Update README.md
Following the Great Recession of 2007-2009, the government established the Consumer Financial Protection Board, CFPB, to police banking practices and protect consumers of financial products from the abuses encountered leading up to the crisis. As part of this, the agency set up a complaint database.
The commplaints come in the form of specially-formatted CSV files.
Using the Python CSV library, this program parses the complaint CSV files, allowing for search, turtle-based-visualization, and various statistics.